requirements.txt
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
docker==5.0.3
Flask==2.0.2
Flask-SocketIO==5.1.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-engineio==4.3.0
python-socketio==5.5.0
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.3
Werkzeug==2.0.2
